add column statuses treeselect
